+++ Editor/Tests/Engine/RumorTest.cs
@@ -19,7 +19,7 @@ namespace Exodrifter.Rumor.Test.Engine
 		{
 			var rumor = new Rumor.Engine.Rumor(new List<Node>());
 
-			var iter = rumor.Run();
+			var iter = rumor.Start();
 			iter.MoveNext();
 
 			Assert.True(rumor.Started);
@@ -40,7 +40,7 @@ namespace Exodrifter.Rumor.Test.Engine
 				new Say("b"),
 			});
 
-			var iter = rumor.Run();
+			var iter = rumor.Start();
 
 			Assert.False(rumor.Started);
 			Assert.False(rumor.Finished);
@@ -96,7 +96,7 @@ namespace Exodrifter.Rumor.Test.Engine
 			Assert.IsFalse(rumor.Running);
 			Assert.IsFalse(rumor.Finished);
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("b", (rumor.Current as Say).EvaluateText(rumor));
 			Assert.IsTrue(rumor.Started);
@@ -122,7 +122,7 @@ namespace Exodrifter.Rumor.Test.Engine
 			Assert.IsFalse(rumor.Running);
 			Assert.IsFalse(rumor.Finished);
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("b", (rumor.Current as Say).EvaluateText(rumor));
 			Assert.IsTrue(rumor.Started);