+++ Editor/Tests/Nodes/CallTest.cs
@@ -22,7 +22,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				new Call("start"),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			Assert.Throws<InvalidOperationException>(() => yield.MoveNext());
 		}
 
@@ -44,7 +44,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				new Call("b"),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("a", (rumor.Current as Say).EvaluateText(rumor));
 
@@ -88,7 +88,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				new Call("a"),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("aa", (rumor.Current as Say).EvaluateText(rumor));
 		}
@@ -114,7 +114,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				}),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("a", (rumor.Current as Say).EvaluateText(rumor));