+++ Editor/Tests/Nodes/LabelTest.cs
@@ -24,7 +24,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				new Say("b"),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("a", (rumor.Current as Say).EvaluateText(rumor));
 		}
@@ -45,7 +45,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				new Say("c"),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("a", (rumor.Current as Say).EvaluateText(rumor));
 
@@ -75,7 +75,7 @@ namespace Exodrifter.Rumor.Test.Nodes
 				new Say("2"),
 			});
 
-			var yield = rumor.Run();
+			var yield = rumor.Start();
 			yield.MoveNext();
 			Assert.AreEqual("1", (rumor.Current as Say).EvaluateText(rumor));